feat: 배팅 시 코인 변경 기능 구현

+++ MC-server.GameRoom/Service/UserTcpService.cs
@@ -26,7 +26,7 @@ namespace MC_server.GameRoom.Service
             {
                 case "coins":
                     // 코인 업데이트
-                    if (value is long coinAmount)
+                    if (value is int coinAmount)
                     {
                         user.Coins += coinAmount;
                         Console.WriteLine($"[socket] Updated coins for user to {user.Coins}");
